@@ -1,10 +1,10 @@
-import { Routes } from '@angular/router';
+import {Routes} from '@angular/router';
 
-import { HomeComponent } from './home/home/home.component';
-import { NotFoundComponent } from './home/not-found/not-found.component';
+import {HomeComponent} from './home/home/home.component';
+import {NotFoundComponent} from './home/not-found/not-found.component';
 
 export const routes: Routes = [
-  { path: '', component: HomeComponent },
+  {path: '', component: HomeComponent},
   {
     path: 'users',
     loadChildren: () => import('./lazy-loaded/users/users.routes').then((m) => m.USERS_ROUTES),
@@ -17,5 +17,5 @@ export const routes: Routes = [
     path: 'lib-dev',
     loadChildren: () => import('./lazy-loaded/lib-dev/lib-dev.routes').then((m) => m.LIB_DEV_ROUTES),
   },
-  { path: '**', component: NotFoundComponent },
+  {path: '**', component: NotFoundComponent},
 ];
