@@ -1,8 +1,8 @@
-import { TestBed } from '@angular/core/testing';
-import { provideRouter } from '@angular/router';
-import { TokenService } from '@auth/services/token.service';
+import {TestBed} from '@angular/core/testing';
+import {provideRouter} from '@angular/router';
+import {TokenService} from '@auth/services/token.service';
 
-import { AuthGuard } from './auth.guard';
+import {AuthGuard} from './auth.guard';
 
 describe('AuthGuard', () => {
   let guard: AuthGuard;
@@ -14,8 +14,8 @@ describe('AuthGuard', () => {
     TestBed.configureTestingModule({
       providers: [
         provideRouter([]),
-        { provide: TokenService, useValue: spyTokenService },
-      ]
+        {provide: TokenService, useValue: spyTokenService},
+      ],
     });
 
     guard = TestBed.inject(AuthGuard);
