@@ -1,22 +1,21 @@
 import {
-  ChangeDetectionStrategy, Component, inject, OnInit
+  ChangeDetectionStrategy, Component, inject, OnInit,
 } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
-import { MAT_DIALOG_DATA, MatDialogModule, MatDialogRef } from '@angular/material/dialog';
-import { AuthService } from '@auth/services/auth.service';
-import { UserLogin } from '@ngnestpostgres/fe-shared';
+import {MatButtonModule} from '@angular/material/button';
+import {MAT_DIALOG_DATA, MatDialogModule, MatDialogRef} from '@angular/material/dialog';
+import {AuthService} from '@auth/services/auth.service';
+import {UserLogin} from '@ngnestpostgres/fe-shared';
 // import { UserRole } from '@ngnestpostgres/fe-shared';
 
 @Component({
   selector: 'anp-login-dialog',
   templateUrl: './login-dialog.component.html',
   styleUrls: ['./login-dialog.component.scss'],
-  standalone: true,
   changeDetection: ChangeDetectionStrategy.OnPush,
   imports: [
     MatButtonModule,
     MatDialogModule,
-  ]
+  ],
 })
 export class LoginDialogComponent implements OnInit {
   private authService = inject(AuthService);
