@@ -1,13 +1,13 @@
-import { TestBed } from '@angular/core/testing';
+import {TestBed} from '@angular/core/testing';
 
-import { TokenService } from './token.service';
+import {TokenService} from './token.service';
 
 describe('TokenService', () => {
   let service: TokenService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      providers: []
+      providers: [],
     });
 
     service = TestBed.inject(TokenService);
