@@ -1,16 +1,16 @@
-import { Inject, Injectable } from '@angular/core';
-import { LocalStorageService } from '@core/services/local-storage.service';
+import {Inject, Injectable} from '@angular/core';
+import {LocalStorageService} from '@core/services/local-storage.service';
 
 const ACCESS_TOKEN = 'access_token';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const REFRESH_TOKEN = 'refresh_token';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class TokenService {
   constructor(
-    @Inject(LocalStorageService) private localStorage: Storage
+    @Inject(LocalStorageService) private localStorage: Storage,
   ) { }
 
   public getAccessToken(): string {
