@@ -1,12 +1,12 @@
-import { HTTP_INTERCEPTORS } from '@angular/common/http';
-import { ClassProvider } from '@angular/core';
+import {HTTP_INTERCEPTORS} from '@angular/common/http';
+import {ClassProvider} from '@angular/core';
 
-import { AuthInterceptor } from './auth.interceptor';
+import {AuthInterceptor} from './auth.interceptor';
 
 export const interceptors: ClassProvider[] = [
   {
     provide: HTTP_INTERCEPTORS,
     multi: true,
-    useClass: AuthInterceptor
-  }
+    useClass: AuthInterceptor,
+  },
 ];
