@@ -1,7 +1,7 @@
-import { TestBed } from '@angular/core/testing';
-import { MatDialogModule } from '@angular/material/dialog';
+import {TestBed} from '@angular/core/testing';
+import {MatDialogModule} from '@angular/material/dialog';
 
-import { DialogsService } from './dialogs.service';
+import {DialogsService} from './dialogs.service';
 
 describe('DialogsService', () => {
   let service: DialogsService;
