@@ -1,11 +1,11 @@
-import { Inject, Injectable, InjectionToken } from '@angular/core';
+import {Inject, Injectable, InjectionToken} from '@angular/core';
 
 export const LOCAL_STORAGE = new InjectionToken<Storage>('Local Storage', {
   providedIn: 'root',
   factory: () => localStorage,
 });
 
-@Injectable({ providedIn: 'root' })
+@Injectable({providedIn: 'root'})
 export class LocalStorageService {
   constructor(@Inject(LOCAL_STORAGE) public storage: Storage) {}
 
