@@ -1,14 +1,14 @@
-import { AsyncPipe } from '@angular/common';
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, ReactiveFormsModule } from '@angular/forms';
-import { MatAutocompleteModule } from '@angular/material/autocomplete';
-import { MatButtonModule } from '@angular/material/button';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatIconModule } from '@angular/material/icon';
-import { MatInputModule } from '@angular/material/input';
-import { Router, RouterModule } from '@angular/router';
+import {AsyncPipe} from '@angular/common';
+import {ChangeDetectionStrategy, Component, OnInit} from '@angular/core';
+import {FormBuilder, FormGroup, ReactiveFormsModule} from '@angular/forms';
+import {MatAutocompleteModule} from '@angular/material/autocomplete';
+import {MatButtonModule} from '@angular/material/button';
+import {MatFormFieldModule} from '@angular/material/form-field';
+import {MatIconModule} from '@angular/material/icon';
+import {MatInputModule} from '@angular/material/input';
+import {Router, RouterModule} from '@angular/router';
 import {
-  debounceTime, distinctUntilChanged, Observable, of, Subject, switchMap
+  debounceTime, distinctUntilChanged, Observable, of, Subject, switchMap,
 } from 'rxjs';
 
 interface SearchOption {
@@ -17,16 +17,15 @@ interface SearchOption {
 }
 
 const OPTIONS: SearchOption[] = [
-  { name: 'Coffees1', link: 'coffees' },
-  { name: 'Users1', link: 'users' },
-  { name: 'Lib1', link: 'lib' },
+  {name: 'Coffees1', link: 'coffees'},
+  {name: 'Users1', link: 'users'},
+  {name: 'Lib1', link: 'lib'},
 ];
 
 @Component({
   selector: 'anp-search-box',
   templateUrl: './search-box.component.html',
   styleUrls: ['./search-box.component.scss'],
-  standalone: true,
   imports: [
     AsyncPipe,
     MatAutocompleteModule,
@@ -35,7 +34,7 @@ const OPTIONS: SearchOption[] = [
     MatIconModule,
     MatInputModule,
     ReactiveFormsModule,
-    RouterModule
+    RouterModule,
   ],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
@@ -86,7 +85,7 @@ export class SearchBoxComponent implements OnInit {
           }
 
           return of<SearchOption[]>(this.searchOptions(value));
-        })
+        }),
       );
   }
 }
