@@ -1,13 +1,13 @@
-import { TestBed } from '@angular/core/testing';
+import {TestBed} from '@angular/core/testing';
 
-import { ThemeManagerService } from './theme-manager.service';
+import {ThemeManagerService} from './theme-manager.service';
 
 describe('ThemeManagerService', () => {
   let service: ThemeManagerService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      providers: []
+      providers: [],
     });
 
     service = TestBed.inject(ThemeManagerService);
