@@ -1,7 +1,7 @@
-import { Routes } from '@angular/router';
+import {Routes} from '@angular/router';
 
-import { CoffeesComponent } from './coffees.component';
+import {CoffeesComponent} from './coffees.component';
 
 export const COFFEES_ROUTES: Routes = [
-  { path: '', component: CoffeesComponent }
+  {path: '', component: CoffeesComponent},
 ];
