@@ -1,8 +1,8 @@
-import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
-import { HttpTestingController, provideHttpClientTesting } from '@angular/common/http/testing';
-import { TestBed } from '@angular/core/testing';
+import {provideHttpClient, withInterceptorsFromDi} from '@angular/common/http';
+import {HttpTestingController, provideHttpClientTesting} from '@angular/common/http/testing';
+import {TestBed} from '@angular/core/testing';
 
-import { CoffeeService } from './coffee.service';
+import {CoffeeService} from './coffee.service';
 
 describe('CoffeeService', () => {
   let service: CoffeeService;
@@ -14,8 +14,8 @@ describe('CoffeeService', () => {
       providers: [
         CoffeeService,
         provideHttpClient(withInterceptorsFromDi()),
-        provideHttpClientTesting()
-      ]
+        provideHttpClientTesting(),
+      ],
     });
 
     service = TestBed.inject(CoffeeService);
