@@ -1,13 +1,13 @@
 import {
-  animate, state, style, transition, trigger
+  animate, state, style, transition, trigger,
 } from '@angular/animations';
 import {
   ChangeDetectionStrategy,
-  Component, EventEmitter, Input, OnInit, Output
+  Component, EventEmitter, Input, OnInit, Output,
 } from '@angular/core';
 
-import { CoffeesFilter } from '../interfaces/coffees-filter';
-import { CoffeeService } from '../services/coffee.service';
+import {CoffeesFilter} from '../interfaces/coffees-filter';
+import {CoffeeService} from '../services/coffee.service';
 
 @Component({
   selector: 'anp-table',
@@ -15,13 +15,12 @@ import { CoffeeService } from '../services/coffee.service';
   styleUrls: ['./table.component.scss'],
   animations: [
     trigger('detailExpand', [
-      state('collapsed', style({ height: '0px', minHeight: '0' })),
-      state('expanded', style({ height: '*' })),
+      state('collapsed', style({height: '0px', minHeight: '0'})),
+      state('expanded', style({height: '*'})),
       transition('expanded <=> collapsed', animate('225ms cubic-bezier(0.4, 0.0, 0.2, 1)')),
     ]),
   ],
-  standalone: true,
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class TableComponent implements OnInit {
   @Input() filters: CoffeesFilter | undefined;
