@@ -1,7 +1,7 @@
-import { Routes } from '@angular/router';
+import {Routes} from '@angular/router';
 
-import { LibDevComponent } from './lib-dev.component';
-import { PhoneFormFieldHostComponent } from './phone-form-field-host/phone-form-field-host.component';
+import {LibDevComponent} from './lib-dev.component';
+import {PhoneFormFieldHostComponent} from './phone-form-field-host/phone-form-field-host.component';
 
 export const LIB_DEV_ROUTES: Routes = [
   {
@@ -10,8 +10,8 @@ export const LIB_DEV_ROUTES: Routes = [
     children: [
       {
         path: 'phone-form-field',
-        component: PhoneFormFieldHostComponent
+        component: PhoneFormFieldHostComponent,
       },
-    ]
-  }
+    ],
+  },
 ];
