@@ -1,38 +1,37 @@
-import { Component } from '@angular/core';
-import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+import {Component} from '@angular/core';
+import {takeUntilDestroyed} from '@angular/core/rxjs-interop';
 import {
-  FormBuilder, FormGroup, ReactiveFormsModule, Validators
+  FormBuilder, FormGroup, ReactiveFormsModule, Validators,
 } from '@angular/forms';
-import { MatButtonModule } from '@angular/material/button';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatIconModule } from '@angular/material/icon';
-import { MatInputModule } from '@angular/material/input';
-import { PhoneFormFieldComponent } from '@ngx-shared';
-import { PhoneParts } from 'projects/ngx-shared/src/public-api';
-import { merge } from 'rxjs';
+import {MatButtonModule} from '@angular/material/button';
+import {MatFormFieldModule} from '@angular/material/form-field';
+import {MatIconModule} from '@angular/material/icon';
+import {MatInputModule} from '@angular/material/input';
+import {PhoneFormFieldComponent} from '@ngx-shared';
+import {PhoneParts} from 'projects/ngx-shared/src/public-api';
+import {merge} from 'rxjs';
 
 @Component({
   selector: 'anp-phone-form-field-host',
   templateUrl: './phone-form-field-host.component.html',
   styleUrls: ['./phone-form-field-host.component.scss'],
-  standalone: true,
   imports: [
     MatButtonModule,
     MatFormFieldModule,
     MatIconModule,
     MatInputModule,
     PhoneFormFieldComponent,
-    ReactiveFormsModule
-  ]
+    ReactiveFormsModule,
+  ],
 })
 export class PhoneFormFieldHostComponent {
   public form!: FormGroup;
 
   constructor(private fb: FormBuilder) {
     this.form = this.fb.group({
-      tel: [{ value: new PhoneParts('', '', ''), disabled: true }, [Validators.required]],
+      tel: [{value: new PhoneParts('', '', ''), disabled: true}, [Validators.required]],
       tel1: [new PhoneParts('+38', '', ''), [Validators.required]],
-      simpleInput: [{ value: '', disabled: true }, [Validators.required]],
+      simpleInput: [{value: '', disabled: true}, [Validators.required]],
       simpleInput1: ['', [Validators.required]],
     });
 
