@@ -1,19 +1,18 @@
-import { Component } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
-import { DialogsService } from '@core/services/dialogs.service';
+import {Component} from '@angular/core';
+import {MatButtonModule} from '@angular/material/button';
+import {DialogsService} from '@core/services/dialogs.service';
 
 @Component({
   selector: 'anp-users-list',
   templateUrl: './users-list.component.html',
   styleUrls: ['./users-list.component.scss'],
-  standalone: true,
-  imports: [MatButtonModule]
+  imports: [MatButtonModule],
 })
 export class UsersListComponent {
   public token: string = 'no token';
 
   constructor(
-    private dialogs: DialogsService
+    private dialogs: DialogsService,
   ) {}
 
   public showLoginForm(): void {
