@@ -1,16 +1,15 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { RouterOutlet } from '@angular/router';
+import {ChangeDetectionStrategy, Component} from '@angular/core';
+import {RouterOutlet} from '@angular/router';
 
 @Component({
   selector: 'anp-users',
-  standalone: true,
   imports: [RouterOutlet],
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: `
-  <div>
-    <router-outlet></router-outlet>
-  </div>
-`,
+    <div>
+      <router-outlet></router-outlet>
+    </div>
+  `,
 })
 export class UsersComponent {
 
