@@ -1,4 +1,4 @@
 export const environment = {
   apiHost: 'https://dev-be-dot-luv-coffee.ew.r.appspot.com/api',
-  production: true
+  production: true,
 };
